Crate cucoqu

cucoqu.rlib — convert cubic, conic, and quadratic bezier segments between one another

TODO: arbitrarily raise degree of an n-degree bezier segment replaces a lot of needlessly used Rust-unsafe Skia code

Modules

co2qu

co2qu.rs — rational (“conic”) quadratic Bézier to quadratic Bézier

cu2qu

cu2qu.rs — cubic Bézier to quadratic Bézier

qu2cu

qu2cu.rs — quadratic to cubic Bézier

Macros

point

Convenience point constructor.

Structs

Point

A two-dimensional point (with x and y coordinates).

Point3

A 3D point type, containing X, Y, and Z coordinates.

Traits

DerivativeCoefficients

The trait for types which implement the calculation of their derivative coefficients.

Point3Like

Type Definitions

Cubic

A cubic Bezier segment.

CubicSpline

A cubic spline.

Quad

A quadratic Bezier segment.

QuadSpline

A quadratic spline.