Trait cucoqu::cu2qu::CubicApprox

pub trait CubicApprox {
    fn approx_control(&self, t: f32) -> Point;
    fn approx_quadratic(
        &self,
        tolerance: f32
    ) -> Result<Quad, ApproxNotFoundError>;
    fn approx_spline(
        &self,
        n: usize,
        tolerance: f32
    ) -> Result<QuadSpline, ApproxNotFoundError>;
}

Cubic Bézier approximation with N quadratic splines

Required Methods

fn approx_control(&self, t: f32) -> Point

Calculates the approximate control point of a cubic Bézier curve to a quadratic Bézier.

fn approx_quadratic(&self, tolerance: f32) -> Result<Quad, ApproxNotFoundError>

Returns a quadratic representation with a tolerance error.

fn approx_spline(
&self,
n: usize,
tolerance: f32
) -> Result<QuadSpline, ApproxNotFoundError>

Returns an array of quadratic spline segments with a tolerance error.